클러스터 컴파일

+++ error.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   error.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: junhalee <junhalee@student.42.fr>          +#+  +:+       +#+        */
+/*   By: junhalee <junhalee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 10:08:42 by junhalee          #+#    #+#             */
-/*   Updated: 2022/01/23 14:04:29 by junhalee         ###   ########.fr       */
+/*   Updated: 2022/01/23 16:19:29 by junhalee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@ void	exit_error(char	*str)
 {
 	if (str)
 		ft_putstr_fd(str, STDERR_FILENO);
-	ft_putendl_fd(strerror(errno), STDERR_FILENO);
+	ft_putendl_fd(strerror((int)errno), STDERR_FILENO);
 	exit(1);
 }
